requirements/base.txt
fastapi==0.92.0
uvicorn==0.20.0
python-dotenv==1.0.0
alembic==1.9.4
mangum==0.17.0
aiohttp==3.8.4
python-amazon-sp-api==0.18.3

sqlalchemy==2.0.4
psycopg2-binary==2.9.5